R/calculate_intensity.R: First version of the groups object specific apply functions
The list object of MetaFeatures have 3 levels:
 1- groups
 2- experiments
 3- bamFiles

A group can contain 1 to n experiments. An experiment can contain 1 to n bam files.

applyOnGroups: To apply a function to every group in the list
applyOnExperiment: To apply a funciton to every experiment in the list
applyOnBamFiles: To apply a function to every bam files in the list

Signed-off-by: Charles Joly Beauparlant <Charles.Joly-Beauparlant@crchul.ulaval.ca>