getGenes no longer call biomaRt (fixes #23)
Getting annotations from biomaRt could take a lot of time.

To avoid this, I saved a compressed object for human and mouse
annotations in the inst/extdata folder:
 * TSS.human
 * TSS.mouse

The objects were generated with the new function getGenesBiomart,
which is essentially the old version of getGenes, but that returns
a GRanges instead of a data.frame.

Signed-off-by: Charles Joly Beauparlant <Charles.Joly-Beauparlant@crchul.ulaval.ca>