Added a bootstrap stat class.
The Bootstrap_Stat is a class that inherit from Basic_Stat. It uses 2 extra
parameters for initialization: sample_size and sample_count.

sample_size is the number of element to pick on each draw during bootstrap. Default
is the size of the input matrix.

sample_count is the number of draw to generate.

Signed-off-by: Charles Joly Beauparlant <charles.joly-beauparlant@crchul.ulaval.ca>