Removed duplicated alignments.
There was a bug where the readGAligments call could import the same read
multiple time.

This is caused by an unexpected behavior of the ScanBamParam used with
readGAligments where a single read can be imported multiple time if the
GRanges used as the which argument of ScanBamParam contains overlapping
regions.

Signed-off-by: Charles Joly Beauparlant <Charles.Joly-Beauparlant@crchul.ulaval.ca>