No more negative coverages (closes #28)
There was a bug where get_normalized_coverage returned negative values.
This was happening when the base coverage was at 0 and the weigth
argument of the GenomicAlignments::coverage function was a numeric with
multiple decimals.

This was caused by an unexpected behavior of the weigth argument that is
relatively unpredictable when not using an integer.

Signed-off-by: Charles Joly Beauparlant <Charles.Joly-Beauparlant@crchul.ulaval.ca>