Changed how coverages are handled.
During metagene initialization, the coverages are no longer splitted by
regions. This means that instead of having as many coverages as there is
combination of bam files / regions, we now have only one coverage by bam
file.

The regions are combined and reduced to target only the relevant
coverages.

This also changed the way matrices are generated since we start with a
smaller number of coverages. The selected approach uses Views which
should greatly reduce the memory usage and the computation time (roughly
2500 regions processed / second during the production of matrices).

The coverages are publicly available through using $coverages on a
metagene object.

The main usage of metagene (first calling new and then plot) will not be
directly affected by this change, meaning that the same command can be
used to produce the same results except for the bin_size argument that
now requires the regions to have the same width (since it does not make
sense to specify a bin size if we need to scale the regions afterward to
produce the matrices.

A new argument is also available for the plot function: bin_count. This
param will work even if the regions have different width.

Signed-off-by: Charles Joly Beauparlant <Charles.Joly-Beauparlant@crchul.ulaval.ca>