Added support for named bam files (closes #50)
The main goal of this commit is to allow the use of a name in the design
files instead of having to use the same file path that was used for the
creation of the metagene object.

For every function that requires a bam name, it is now possible to use
the file name with path (actually, the path will be ignored and only
the basename without the extension will be used), the file name without
the path or the name of the bam file if a named bam file vector was used
to create the metagene object.

Signed-off-by: Charles Joly Beauparlant <charles.joly-beauparlant@crchul.ulaval.ca>