Improved the metagene$plot interface (closes #42).
Added the possibility to chose between the "bootstrap" and "basic" stats
to calculate the values for the ribbon. For this purpose, there is now
the "stat" parameter. Extra params specific for the statistical
calculation will now be passed through the "..." parameter. These params
include "aplha", "average", "sample_count" and "sample_size". See
?metagene for a description of each of those parameters.

Signed-off-by: Charles Joly Beauparlant <Charles.Joly-Beauparlant@crchul.ulaval.ca>