Extracted the plot_metagene function.
The function that produced the metagene (i.e.: that called ggplot) was
in the private section of the metagene object. It is now an exported
function called `plot_metagene`.

The goal is to be able to combine `data.frame`s produced by different
metagene analysis and to be able to plot the result without having to
launch a new metagene analysis.

Signed-off-by: Charles Joly Beauparlant <charles.joly-beauparlant@crchul.ulaval.ca>